% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{$,dgeComparison-method}
\alias{$,dgeComparison-method}
\title{Returns a column of the DE results data frame.}
\usage{
\S4method{$}{dgeComparison}(x, name)
}
\arguments{
\item{x}{An object of class \code{dgeComparison}}
}
\value{
A column of the DE results data frame.
}
\description{
Perhaps think about making it convenient to also return metadata this way??
}
