% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\docType{class}
\name{dgeComparison-class}
\alias{dgeComparison-class}
\title{The \code{dgeComparison} class}
\description{
The basic class holds the DE results for a particular comparison along with meta-data.
The meta-data is stored in specific slots and getter methods can be used to access them.
The DE results assume output from limma-voom and calculate effect sizes and their variances
when the object is created. Methods are present to subset this object as well. Subsetting the object
means subsetting the DE results only; all the meta-data stays the same.
}
\section{Slots}{

\describe{
\item{\code{deRes}}{A data.frame which contains the DE results. Has the following columns:
\enumerate{
  \item id: internal gene id. Usually a concatenation of mouse and rat Ensembl
  \item ensembl: Ensemble id
  \item symbol: gene symbol
  \item logFC: log fold-change
  \item se: standard error
  \item t: t-statistic
  \item pvalue: associated pvalue
  \item df.residual: residual degrees of freedom
  \item df.total: total degrees of freedom
  \item cohensD: effect size statistic Cohen's D
  \item varD: variance associated with Cohen's D
  \item hedgesG: effect size statistic Hedge's G
  \item varG: variance associated with Hedge's G
}}

\item{\code{name}}{The internal name of the experiment that this comparison is from}

\item{\code{model}}{The model of the experiment. Possible values include: GRov, FSL, HR-LR, CSDS, Cort, Stress}

\item{\code{treatment}}{If any treatments have been used in this particular comparison.}

\item{\code{housing}}{If enriched or standard housing was used}

\item{\code{timepoint}}{The timepoint for an experiment with various timepoints.}

\item{\code{species}}{The species include mouse, rat, human}

\item{\code{region}}{The brain region. possible values include HPC, vHPC, dHPC, DG, vDG, dDG, PFC, NACC, STR}

\item{\code{sex}}{Female or Male}

\item{\code{platform}}{MicroArray or RNASeq}

\item{\code{group1}}{The first group name. The comparison ALWAYS givew the results for Group1 - Group2}

\item{\code{group2}}{The second group name}

\item{\code{n.group1}}{Group 1 N}

\item{\code{n.group2}}{GRoup 2 N}

\item{\code{n.effective}}{Internally calculated in order to calculate effect sizes}

\item{\code{comparisonID}}{The internal comparison ID.}
}}

