% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\docType{class}
\name{dgeData-class}
\alias{dgeData-class}
\title{Class `dgeData`}
\description{
Has only one slot, a list with all the experiments that form this DGE data set.
}
\section{Slots}{

\describe{
\item{\code{experiments}}{list of `dgeExperiments`}

\item{\code{ids}}{A data.table of unique DGE ids present in this data set along with the species ensembl ids and symbols}
}}

