% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{dgeData}
\alias{dgeData}
\title{Constructor for the `dgeData` class. Creates the dataset ids when called.}
\usage{
dgeData(experiments)
}
\arguments{
\item{experiments}{The list of experiments. Must be of class `dgeExperiment`.}
}
\value{
A new `dgeData` object.
}
\description{
Constructor for the `dgeData` class. Creates the dataset ids when called.
}
