% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{flattenDge,dgeComparison-method}
\alias{flattenDge,dgeComparison-method}
\title{Returns a data.table with all relevant information}
\usage{
\S4method{flattenDge}{dgeComparison}(o, geneId = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeComparison} from which to get the data}

\item{geneId}{An optional vector of geneId for which to get the information}
}
\value{
A data.table with all relevant information
}
\description{
Returns a data.table with all relevant information
}
