% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{flattenDge,dgeData-method}
\alias{flattenDge,dgeData-method}
\title{Flatten a data set to a data table}
\usage{
\S4method{flattenDge}{dgeData}(o, geneId = NULL, conn = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeData}}

\item{geneId}{Information for only these id's is returned}

\item{conn}{A connection to an SQLite DB. Has to be of class \code{SQLiteConnection}.}
}
\value{
A data.table conatining all the information in the experiment.
}
\description{
Flatten a data set to a data table
}
