% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{flattenDge,dgeExperiment-method}
\alias{flattenDge,dgeExperiment-method}
\title{Flatten an experiment to a data table}
\usage{
\S4method{flattenDge}{dgeExperiment}(o, geneId = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}

\item{geneId}{Information for only these id's is returned}
}
\value{
A data.table conatining all the information in the experiment.
}
\description{
Flatten an experiment to a data table
}
