% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{geneId,dgeData-method}
\alias{geneId,dgeData-method}
\title{Get ids for a data set}
\usage{
\S4method{geneId}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
a character vector of internal DGE ids
}
\description{
Get ids for a data set
}
