% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{geneId,dgeExperiment-method}
\alias{geneId,dgeExperiment-method}
\title{Get the gene DGE id's for this experiment}
\usage{
\S4method{geneId}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector with DGE id's
}
\description{
Get the gene DGE id's for this experiment
}
\examples{
geneId(dgeExp)
}
