% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getAllComparisons,dgeExperiment-method}
\alias{getAllComparisons,dgeExperiment-method}
\title{Get all comparisons}
\usage{
\S4method{getAllComparisons}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A list of objects of class `dgeComparison`.
}
\description{
Get all comparisons
}
