% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getColumnData,dgeData-method}
\alias{getColumnData,dgeData-method}
\title{Get column data for particular comparisons}
\usage{
\S4method{getColumnData}{dgeData}(o, column, comparisonID, conn = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeData}}

\item{column}{This can be any column name in the \code{data.table} returned by \code{flattenDge}. These include
\enumerate{
    \item logFC
    \item se
    \item pvalue
    \item fdr
    \item cohensD
    \item varD
    \item hedgesG
    \item varG
}}

\item{comparisonID}{The comparison IDs for which the data is needed}

\item{conn}{The connection to the SQLite DB if the data is stored there}
}
\value{
A data.table with columns as comparisonIDs
}
\description{
Get column data for particular comparisons
}
