% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getComparison,dgeExperiment-method}
\alias{getComparison,dgeExperiment-method}
\title{Get one or many comparisons from an experiment object}
\usage{
\S4method{getComparison}{dgeExperiment}(
  o,
  groups = NULL,
  treatment = NULL,
  timepoint = NULL,
  region = NULL,
  sex = NULL
)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment} from which the comparison is to be extracted}

\item{groups}{The groups for which the comparison is needed. If specified, the other two parameters are ignored}

\item{treatment}{Optional parameter specifying the treatments needed for the comparison. Can be used in conjunction with others}

\item{timepoint}{Optional parameter specifying the timepoints needed for the comparison. Can be used in conjunction with others}

\item{region}{The brain region of the desired comparison. Can be specified in conjunction with treatment}

\item{sex}{The sex of the desired comparison}
}
\value{
If no comparison is matched, `NULL` is returned, else an object of class `dgeExperiment` with the specified comparisons.
}
\description{
The comparisons matching the given parameters are returned from the experiment. If groups are specified, they need to be a
character vector of length two. If need be, the comparison will be reversed. One of groups, treatment or region
needs to be specified. If both treatment and region are specified, the comparison will be matched using both parameters.
In general, this performs an AND, not an OR. Comparisons matching ALL parameters will be returned.
}
\examples{
#getComparison(fslExperiment, groups = c("FSL", "FRL"))
#getComparison(fslExperiment, region = "HPC")
}
