% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getComparisonById,dgeData-method}
\alias{getComparisonById,dgeData-method}
\title{Gets the comparisons within this data set by id.}
\usage{
\S4method{getComparisonById}{dgeData}(o, id)
}
\arguments{
\item{o}{An object of class \code{dgeData} which we want to subset}

\item{id}{A character vector with the ids we need}
}
\value{
An \code{dgeData} object with the needed ids. If none found, NULL is returned.
}
\description{
Gets the comparisons within this data set by id.
}
