% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getComparisonById,dgeExperiment-method}
\alias{getComparisonById,dgeExperiment-method}
\title{Gets the comparisons within this experiment by id.}
\usage{
\S4method{getComparisonById}{dgeExperiment}(o, id)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment} which we want to subset}

\item{id}{A character vector with the ids we need}
}
\value{
An \code{dgeExperiment} object with the needed ids. If none found, NULL is returned.
}
\description{
Gets the comparisons within this experiment by id.
}
