% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getDERes,dgeComparison-method}
\alias{getDERes,dgeComparison-method}
\title{Gets the DE Results}
\usage{
\S4method{getDERes}{dgeComparison}(o, groups = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeComparison} from which to return the DE result}

\item{groups}{An optional groups vector. If the groups don't correspond to the comparison groups, an error is thrown.
However, if the groups are present and correct and the order of the groups is reversed, the DE results returned are reversed also.}
}
\value{
A data frame containing the DE results and effect sizes for all genes.
}
\description{
This is a data frame containing the per gene DE results for this particular comparison.
}
