% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getExperiments,dgeData-method}
\alias{getExperiments,dgeData-method}
\title{Get all experiments in a data set.}
\usage{
\S4method{getExperiments}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A list of objects of class \code{dgeExperiment}
}
\description{
Get all experiments in a data set.
}
