% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getExperimentsBySpecies,dgeData-method}
\alias{getExperimentsBySpecies,dgeData-method}
\title{Gets the experiments of a particular species in an DGE data set}
\usage{
\S4method{getExperimentsBySpecies}{dgeData}(o, species)
}
\arguments{
\item{o}{An object of class \code{dgeData}}

\item{species}{ONE species for which to get the experiments}
}
\value{
\code{NULL} if there are no experiments for that species. Else an object of class \code{dgeData} with experiments from \code{species}
}
\description{
Gets the experiments of a particular species in an DGE data set
}
