% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getGeneEns,dgeComparison-method}
\alias{getGeneEns,dgeComparison-method}
\title{Get a vector of ensembl id's for the genes in this comparison.}
\usage{
\S4method{getGeneEns}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A vector of ensemble ids
}
\description{
Get a vector of ensembl id's for the genes in this comparison.
}
