% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getGeneEns,dgeData-method}
\alias{getGeneEns,dgeData-method}
\title{Get the ensemble id's for a particular species in a DGE data set}
\usage{
\S4method{getGeneEns}{dgeData}(o, species)
}
\arguments{
\item{o}{An object of class \code{dgeData} from which to get id's}

\item{species}{The species for which to get id's. Can be a vector.}
}
\value{
A list with one element for each species giving the unique Ensembl id's present in each experiment.
}
\description{
Get the ensemble id's for a particular species in a DGE data set
}
\examples{
getGeneEnsembl(dataSet, "mouse")
getGeneEnsembl(dataSet, c("Mouse", "Rat"))
}
