% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getGeneSymbol,dgeComparison-method}
\alias{getGeneSymbol,dgeComparison-method}
\title{Get a vector of gene symbols for this comparison.}
\usage{
\S4method{getGeneSymbol}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A vector of gene symbols
}
\description{
Get a vector of gene symbols for this comparison.
}
