% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getGeneSymbol,dgeData-method}
\alias{getGeneSymbol,dgeData-method}
\title{Get the gene symbols for a particular species in a DGE data set}
\usage{
\S4method{getGeneSymbol}{dgeData}(o, species)
}
\arguments{
\item{o}{An object of class \code{dgeData} from which to get the symbols}

\item{species}{The species for which to get the symbols. Can be a vector.}
}
\value{
A list with one element for each species giving the unique gene symbols present in each experiment.
}
\description{
Get the gene symbols for a particular species in a DGE data set
}
\examples{
getGeneEnsembl(dataSet, "mouse")
getGeneEnsembl(dataSet, c("Mouse", "Rat"))
}
