% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getGeneSymbol,dgeExperiment-method}
\alias{getGeneSymbol,dgeExperiment-method}
\title{Get the gene symbols for this experiment}
\usage{
\S4method{getGeneSymbol}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector with gene symbols
}
\description{
Get the gene symbols for this experiment
}
