% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getGroups,dgeComparison-method}
\alias{getGroups,dgeComparison-method}
\title{Gets the groups for this comparison.}
\usage{
\S4method{getGroups}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A names character vector of length two with the group names.
}
\description{
Returns the names of the groups in the comparison. The comparison is always Group1 - Group2
}
\examples{
getGroups(dge)
}
