% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getGroups,dgeData-method}
\alias{getGroups,dgeData-method}
\title{Get a vector of groups}
\usage{
\S4method{getGroups}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A string vector of groups which are a part of the data set.
}
\description{
Get a vector of groups
}
