% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getGroups,dgeExperiment-method}
\alias{getGroups,dgeExperiment-method}
\title{Get a vector of groups}
\usage{
\S4method{getGroups}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A string vector of groups which are a part of the experiment.
}
\description{
A possible complication. Should we ensure that each and every combination of groups has a comparison present?
Don't think we need to make this a requirement, but just make it highly recommended that a two-way comparison
with all group combinations is present.
}
