% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getHousing,dgeComparison-method}
\alias{getHousing,dgeComparison-method}
\title{Get the treatment for this comparison. If no treatment, "None" is returned.}
\usage{
\S4method{getHousing}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with the housing name
}
\description{
Get the treatment for this comparison. If no treatment, "None" is returned.
}
\examples{
getTreatment(dge)
}
