% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getHousing,dgeExperiment-method}
\alias{getHousing,dgeExperiment-method}
\title{Get the housing in the experiment}
\usage{
\S4method{getHousing}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector of the housing.
}
\description{
Get the housing in the experiment
}
\examples{
getTreatments(dgeExperiment)
}
