% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getIds,dgeData-method}
\alias{getIds,dgeData-method}
\title{Get the internal comparison IDs of the comparisons in this data base}
\usage{
\S4method{getIds}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData} from which to get the IDs}
}
\value{
A character vector of IDs
}
\description{
Get the internal comparison IDs of the comparisons in this data base
}
