% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getIds,dgeExperiment-method}
\alias{getIds,dgeExperiment-method}
\title{Get all the internal comparison IDs in this experiment}
\usage{
\S4method{getIds}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment} from which we want the IDs}
}
\value{
A character vector of IDs
}
\description{
Get all the internal comparison IDs in this experiment
}
