% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getModel,dgeComparison-method}
\alias{getModel,dgeComparison-method}
\title{Get the model name}
\usage{
\S4method{getModel}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
the model name
}
\description{
Get the model name
}
