% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getModel,dgeData-method}
\alias{getModel,dgeData-method}
\title{Get the models present in this data set}
\usage{
\S4method{getModel}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector with the model names.
}
\description{
Get the models present in this data set
}
