% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getName,dgeComparison-method}
\alias{getName,dgeComparison-method}
\title{Get the experiment name}
\usage{
\S4method{getName}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
the experiment name
}
\description{
Usually of the form LabName-experimentNumber. Is the internal DGE identifier for this particular experiment.
}
\examples{
getName(dge)
}
