% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getName,dgeData-method}
\alias{getName,dgeData-method}
\title{Get the experiment names present in this data set}
\usage{
\S4method{getName}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector with the experiment names.
}
\description{
Get the experiment names present in this data set
}
