% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getName,dgeExperiment-method}
\alias{getName,dgeExperiment-method}
\title{Returns the experiment name}
\usage{
\S4method{getName}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
The experiment name
}
\description{
Returns the experiment name
}
