% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getOwner,dgeExperiment-method}
\alias{getOwner,dgeExperiment-method}
\title{Returns the lab where this experiment was conducted. If it is private,
only lab members will have access to this data.}
\usage{
\S4method{getOwner}{dgeExperiment}(o)
}
\arguments{
\item{dgeExperiment}{}
}
\value{

}
\description{
Returns the lab where this experiment was conducted. If it is private,
only lab members will have access to this data.
}
