% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getPlatform,dgeData-method}
\alias{getPlatform,dgeData-method}
\title{Get the platforms present in this data set.}
\usage{
\S4method{getPlatform}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector of platforms, either Microarray or RNASeq or both.
}
\description{
Get the platforms present in this data set.
}
