% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getPlatform,dgeExperiment-method}
\alias{getPlatform,dgeExperiment-method}
\title{The platform: microarray or rna-seq}
\usage{
\S4method{getPlatform}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
String specifying the platform
}
\description{
The platform: microarray or rna-seq
}
\examples{
getPlatform(dgeExperiment)
}
