% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getPublic,dgeExperiment-method}
\alias{getPublic,dgeExperiment-method}
\title{Title}
\usage{
\S4method{getPublic}{dgeExperiment}(o)
}
\arguments{
\item{dgeExperiment}{}
}
\value{

}
\description{
Title
}
