% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getRegion,dgeComparison-method}
\alias{getRegion,dgeComparison-method}
\title{Gets the brain region.}
\usage{
\S4method{getRegion}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with brain region.
}
\description{
One of "HPC", "vHPC", "dHPC", "DG", "vDG", "dDG", "PFC", "NACC", "STR" or "VTA".
}
