% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getRegion,dgeData-method}
\alias{getRegion,dgeData-method}
\title{Get the brain regions present in the DGE data set}
\usage{
\S4method{getRegion}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector of brain regions present in this DGE data set.
}
\description{
Get the brain regions present in the DGE data set
}
