% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getSex,dgeComparison-method}
\alias{getSex,dgeComparison-method}
\title{Sex of the animals used in this comparison.}
\usage{
\S4method{getSex}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character vector with the sex.
}
\description{
"male" or "female" or "both"
}
