% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getSex,dgeData-method}
\alias{getSex,dgeData-method}
\title{Get the sexes present in this DGE data set}
\usage{
\S4method{getSex}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector of the sexes which are included in this data set. Can include "both" if some comparisons are across sexes.
}
\description{
Get the sexes present in this DGE data set
}
