% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getSpecies,dgeData-method}
\alias{getSpecies,dgeData-method}
\title{Get Species in this data set}
\usage{
\S4method{getSpecies}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector of species in the data set. Currently limited to "mouse", "rat" and "human"
}
\description{
Get Species in this data set
}
