% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getTimepoint,dgeData-method}
\alias{getTimepoint,dgeData-method}
\title{Get the treatments present in this data set}
\usage{
\S4method{getTimepoint}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData}}
}
\value{
A character vector of the timepoint in this data set.
}
\description{
Get the treatments present in this data set
}
