% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getTreatment,dgeComparison-method}
\alias{getTreatment,dgeComparison-method}
\title{Get the treatment for this comparison. If no treatment, "None" is returned.}
\usage{
\S4method{getTreatment}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with the treatment name
}
\description{
Get the treatment for this comparison. If no treatment, "None" is returned.
}
