% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getTreatment,dgeExperiment-method}
\alias{getTreatment,dgeExperiment-method}
\title{Get the treatments in the experiment}
\usage{
\S4method{getTreatment}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector of the treatments. "None" is returned if there are none.
}
\description{
Get the treatments in the experiment
}
\examples{
getTreatments(dgeExperiment)
}
