% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{names,dgeComparison-method}
\alias{names,dgeComparison-method}
\title{Overloading names to get the name of the comparison}
\usage{
\S4method{names}{dgeComparison}(x)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A string with the name
}
\description{
Overloading names to get the name of the comparison
}
