% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{nrow,dgeComparison-method}
\alias{nrow,dgeComparison-method}
\title{Returns the number of genes in the DE results}
\usage{
\S4method{nrow}{dgeComparison}(x)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Integer with the number of genes.
}
\description{
Returns the number of genes in the DE results
}
\examples{
nrow(dge)
}
