% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{printComparison,dgeExperiment-method}
\alias{printComparison,dgeExperiment-method}
\title{Print the comparisons in this experiment object.}
\usage{
\S4method{printComparison}{dgeExperiment}(
  o,
  name = FALSE,
  region = FALSE,
  treatment = FALSE,
  housing = FALSE,
  timepoint = FALSE
)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment} for which to print the comparisons}

\item{name}{Logical, indicating if the name of the experiment should be printed. Default \code{FALSE}.}

\item{region}{Logical, indicating if the brain region of the comparison should be printed. Default \code{FALSE}.}

\item{treatment}{Logical, indicating if the treatment of the comparison should be printed. Default \code{FALSE}.}

\item{housing}{Logical, indicating if the housing of the comparison should be printed. Default \code{FALSE}.}

\item{timepoint}{Logical, indicating if the timepoint of the comparison should be printed. Default \code{FALSE}.}
}
\value{
A vector of strings with the comparisons
}
\description{
Print the comparisons in this experiment object.
}
