% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{removeComparison,dgeData-method}
\alias{removeComparison,dgeData-method}
\title{Removes the comparisons which match the given id's}
\usage{
\S4method{removeComparison}{dgeData}(o, id)
}
\arguments{
\item{o}{The \code{dgeData} object from which to remove the id's}

\item{id}{A character vector of the ids to remove}
}
\value{
An object of class \code{dgeData} with the id's removed
}
\description{
Removes the comparisons which match the given id's
}
