% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{removeComparison,dgeExperiment-method}
\alias{removeComparison,dgeExperiment-method}
\title{Remove comparisons from an experiment}
\usage{
\S4method{removeComparison}{dgeExperiment}(o, id)
}
\arguments{
\item{o}{The expermient from which to remove the comparisons}

\item{id}{A character vector of the ids of the comparison to remove.}
}
\value{
The experiment with the comparison removed
}
\description{
Remove comparisons from an experiment
}
