% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{reverseComparison,dgeComparison-method}
\alias{reverseComparison,dgeComparison-method}
\title{Convenience method to reverse the direction of a comparison}
\usage{
\S4method{reverseComparison}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}

\item{comparisons}{Ignored for this class}
}
\value{
An object of class dgeComparison
}
\description{
Convenience method to reverse the direction of a comparison
}
