% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{reverseComparison,dgeData-method}
\alias{reverseComparison,dgeData-method}
\title{Reverses a set of comparisons in the data}
\usage{
\S4method{reverseComparison}{dgeData}(o, comparisons)
}
\arguments{
\item{o}{An object of class \code{dgeData}}

\item{comparisons}{A character vector with the list of comparisons to be reversed}
}
\value{

}
\description{
Reverses a set of comparisons in the data
}
