% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{reverseComparison,dgeExperiment-method}
\alias{reverseComparison,dgeExperiment-method}
\title{Reverse comparisons in this experiment}
\usage{
\S4method{reverseComparison}{dgeExperiment}(o, comparisons)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}

\item{comparisons}{A character vector of comparisonIDs which are to be reversed}
}
\value{
The \code{dgeExperiment} object with the IDs reversed
}
\description{
Reverse comparisons in this experiment
}
