% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{show,dgeComparison-method}
\alias{show,dgeComparison-method}
\title{Print an DGE Comparison}
\usage{
\S4method{show}{dgeComparison}(object)
}
\arguments{
\item{dgeComparison}{}
}
\description{
Prints the basic meta-data from an dgeComparison
}
