% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{show,dgeData-method}
\alias{show,dgeData-method}
\title{Print the experiments names in the data set.}
\usage{
\S4method{show}{dgeData}(object)
}
\arguments{
\item{dgeData}{}
}
\description{
Print the experiments names in the data set.
}
