% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{show,dgeExperiment-method}
\alias{show,dgeExperiment-method}
\title{Show method.}
\usage{
\S4method{show}{dgeExperiment}(object)
}
\arguments{
\item{dgeExperiment}{}
}
\value{

}
\description{
Prints the experiment
}
