% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{[,dgeComparison,ANY,ANY,ANY-method}
\alias{[,dgeComparison,ANY,ANY,ANY-method}
\title{Subset the DE results of this comparison.}
\usage{
\S4method{[}{dgeComparison,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{An object of class \code{dgeComparison}}
}
\value{
An dgeComparison object with the specified DE results.
}
\description{
It doesn't effect the meta-data in any way. It mirrors base R's subset functions for a data.frame.
}
