% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{[,dgeData-method}
\alias{[,dgeData-method}
\title{Subset by gene id}
\usage{
\S4method{[}{dgeData}(x, i)
}
\arguments{
\item{x}{An object of class \code{dgeData}}
}
\value{
and object of class `dgeData`
}
\description{
This function returns a subsetted dgeData object containing all the experiments but only the selected genes for which an id is given.
i can be a vector of id's or a logical vector with the same number of rows as the data set.
}
\examples{
`dge[c(TRUE, FALSE, FALSE, TRUE)]`
}
