% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{[,dgeExperiment-method}
\alias{[,dgeExperiment-method}
\title{Subset operator for DGE Experiment}
\usage{
\S4method{[}{dgeExperiment}(x, i)
}
\arguments{
\item{x}{An object of class \code{dgeExperiment}}
}
\value{
A subsetted object of class `dgeExperiment`.
}
\description{
Rather simplistic. Don't specify `j`... it won't do anything. Just `i` as a logical vector will work. It would be
best to use `getGeneEns` to get the Ensembl id's and subset using those.
}
