% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{symbolToId,dgeData-method}
\alias{symbolToId,dgeData-method}
\title{Converts human gene symbols to the internal IDs used by the package.}
\usage{
\S4method{symbolToId}{dgeData}(o, geneSymbols, species = NULL)
}
\arguments{
\item{o}{An object of class \code{dgeData} from which to extract the IDs}

\item{geneSymbols}{The symbols for which we need the IDs}

\item{species}{The species which the symbols correpond to. If unspecified, the function returns matches across all species.}
}
\value{
A character vector with the gene ids which can be used to subset the data
}
\description{
Converts human gene symbols to the internal IDs used by the package.
}
