% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{dgeComparison}
\alias{dgeComparison}
\title{Constructor function for the dgeComparison class.}
\usage{
dgeComparison(
  deData,
  name,
  model,
  treatment = "None",
  housing = "Standard",
  timepoint = "None",
  species,
  region,
  sex,
  platform,
  groups,
  Ns
)
}
\arguments{
\item{deData}{A file name or a data frame containing the DE results. if a file name, it must be a CSV file.
The following columns are expected in the file or data frame:
\itemize{
  \item ensembl: The Ensembl id for the gene
  \item symbol: The gene symbol
  \item logFC: the lof fold change
  \item se: the standard error for the log fold change
  \item t: the t-statistic for the gene
  \item pvalue: the associated pvalue
  \item df.residual: the residual degrees of freedom
  \item df.total: the total degrees of freedom
}}

\item{name}{The name of the experiment.}

\item{model}{The model name.}

\item{treatment}{The treatment, if any. If absent, treatment will be specified as "None".}

\item{species}{The species. One of "mouse", "rat" or "human"}

\item{region}{The brain region}

\item{sex}{The sex. "female", "male" or "both"}

\item{platform}{The experimental platform. MIcroArray or RNASeq.}

\item{groups}{The two groups. A character vector of length two is expected.}

\item{Ns}{The number of samples in each group. Integer vector of length two is expected.}
}
\value{
A dgeComparison object.
}
\description{
The function takes in the DE results or the file name containing the DE results, along with
the meta-data and checks it for any errors. The effect sizes are also calculated at this stage and
the resultant dgeComparison object is returned. The effect sizes calculated are Cohen's D and Hedge's G
along with their variances.
}
