% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\docType{class}
\name{dgeExperiment-class}
\alias{dgeExperiment-class}
\title{The \code{dgeExperiment} class}
\description{
The dgeExperiment class holds a list of dgeComparisons. The comparisons belong to the same
experiment, so have the same name. They should also have the same species and the same model name
and platform. What can change is the treatments, the brain regions and the sex. This class should list ALL
The treatments, regions and sexes involved in this experiment. In fact, this is a requirement when creating
this class from a list of comparisons.
}
\section{Slots}{

\describe{
\item{\code{comparisons}}{A list of objects of the class dgeComparison which comprise this experiment. Needs to be at lest 1.}

\item{\code{name}}{The name of the experiment. Usually lab name with an hyphenated number.}

\item{\code{model}}{The model name.}

\item{\code{species}}{The species.}

\item{\code{platform}}{The platform.}

\item{\code{treatments}}{A character vector with the list of treatments for each comparison. If there aren't any then "None"}

\item{\code{timepoints}}{A character vector with the timepoints for each comparison. If there aren't any, then "None"}

\item{\code{regions}}{A character vector which specifies the brain regions in this particular experiment}

\item{\code{sexes}}{The sex: male or female or both}

\item{\code{groups}}{The different groups in the experiment}

\item{\code{totalN}}{The total number of animals in this experiment}

\item{\code{description}}{Text describing the experiment. This is optional but highly recommended.}

\item{\code{owner}}{The system username of the owner. Only useful if it is a private data set}

\item{\code{public}}{Logical flag. If true, then the data set can be viewed by all with access to the system.}
}}

