% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{dgeExperiment}
\alias{dgeExperiment}
\title{Constructor for the `dgeExperiment` class.}
\usage{
dgeExperiment(comparisons, description = "")
}
\arguments{
\item{comparisons}{A `list` of comparisons belonging to the same experiment}

\item{description}{An optional description of the experiment.}
}
\value{
An object of class`dgeExperiment`
}
\description{
Takes a list of `dgeComparisons` belonging to the same experiment and creates an object of `dgeExperiment` from them.
}
