% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{geneId,dgeComparison-method}
\alias{geneId,dgeComparison-method}
\title{Gets the vector of DGE id's for the genes in this comparison.}
\usage{
\S4method{geneId}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A vector of DGE ids
}
\description{
Gets the vector of DGE id's for the genes in this comparison.
}
\examples{
geneId(dge)
}
