% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{geneId<-,dgeComparison-method}
\alias{geneId<-,dgeComparison-method}
\title{Sets the DGE gene ids for a `dgeComparison`}
\usage{
\S4method{geneId}{dgeComparison}(o) <- value
}
\arguments{
\item{o}{An object of class \code{dgeComparison} for which the values need to be set}

\item{value}{A data.table with ids and ensembl.}
}
\value{
An object of class `dgeComparison` with the ids set
}
\description{
Sets the DGE gene ids for a `dgeComparison`
}
\examples{
geneId(dge) <- ids
}
