% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{geneId<-,dgeExperiment-method}
\alias{geneId<-,dgeExperiment-method}
\title{Set the gene DGE id's for this experiment}
\usage{
\S4method{geneId}{dgeExperiment}(o) <- value
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}

\item{value}{A data.table with ids and ensembl values.}
}
\value{

}
\description{
Set the gene DGE id's for this experiment
}
