% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{getComparison,dgeData-method}
\alias{getComparison,dgeData-method}
\title{Get one or many comparisons from a `dgeData` object.}
\usage{
\S4method{getComparison}{dgeData}(
  o,
  groups = NULL,
  treatment = NULL,
  timepoint = NULL,
  region = NULL,
  sex = NULL,
  species = NULL,
  model = NULL,
  name = NULL
)
}
\arguments{
\item{o}{An object of class \code{dgeData} from which the comparison is to be extracted}

\item{groups}{The groups for which the comparison is needed. If specified, the other two parameters are ignored}

\item{treatment}{Optional parameter specifying the treatment needed for the comparison. Can be used in conjunction with region}

\item{region}{The brain region of the desired comparison. Can be specified in conjunction with treatment}

\item{sex}{The sex of the desired comparison}

\item{species}{The desired species}

\item{model}{Select only specified models}

\item{name}{Select only experiments with specified name}
}
\value{
If no comparison is matched, `NULL` is returned, else an object of class `dgeData` with the specified comparisons.
}
\description{
The comparisons matching the given parameters are returned from the given data set. If groups are specified, they need to be a
character vector of length two. If need be, the comparison will be reversed. One of groups, treatment or region
needs to be specified. If both treatment and region are specified, the comparison will be matched using both parameters.
}
