% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getGeneEns,dgeExperiment-method}
\alias{getGeneEns,dgeExperiment-method}
\title{Get the gene ensembl id's for this experiment}
\usage{
\S4method{getGeneEns}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector with ensembl id's
}
\description{
Get the gene ensembl id's for this experiment
}
\examples{
getGeneEns(dgeExp)
}
