% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getModel,dgeExperiment-method}
\alias{getModel,dgeExperiment-method}
\title{Returns the experiment model}
\usage{
\S4method{getModel}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character with the experiment model
}
\description{
Returns the experiment model
}
