% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getNs,dgeComparison-method}
\alias{getNs,dgeComparison-method}
\title{Gets the number of animals per group.}
\usage{
\S4method{getNs}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
A named integer vector of length two with the group sizes.
}
\description{
Is used to calculate the effect size. Not sure how useful it will be for users, but its there just in case.
}
