% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getOwner,dgeExperiment-method}
\alias{getOwner,dgeExperiment-method}
\title{Returns the lab where this experiment was conducted. If it is private,
only lab members will have access to this data.}
\usage{
\S4method{getOwner}{dgeExperiment}(o)
}
\arguments{
<<<<<<< HEAD
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character string with the owner of the data. Returns \code{NULL} if not set.
=======
\item{dgeExperiment}{}
}
\value{

>>>>>>> 7942bd79adcb41586780467972f5818a4b964d91
}
\description{
Returns the lab where this experiment was conducted. If it is private,
only lab members will have access to this data.
}
