% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getPlatform,dgeComparison-method}
\alias{getPlatform,dgeComparison-method}
\title{Gets the experiment platform.}
\usage{
\S4method{getPlatform}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with the platform name.
}
\description{
Either MicroArray or RNASeq. Maybe extend to others like ChIPSeq?
}
\examples{
getPlatform(dge)
}
