% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getPublic,dgeExperiment-method}
\alias{getPublic,dgeExperiment-method}
\title{Returns a boolean value which is TRUE if the data is universally accessible.}
\usage{
\S4method{getPublic}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A logical value, \code{TRUE} if it is public, \code{FALSE} otherwise.
}
\description{
Returns a boolean value which is TRUE if the data is universally accessible.
}
