% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getRegion,dgeExperiment-method}
\alias{getRegion,dgeExperiment-method}
\title{Get the brain regions of the experiment}
\usage{
\S4method{getRegion}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector of the brain regions.
}
\description{
Get the brain regions of the experiment
}
\examples{
getRegions(dgeExperiment)
}
