% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getSex,dgeExperiment-method}
\alias{getSex,dgeExperiment-method}
\title{Get the sexes in the experiment.}
\usage{
\S4method{getSex}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector of the sexes of which we have comparisons. Can include "both"
}
\description{
Get the sexes in the experiment.
}
\examples{
getSex(dgeExperiment)
}
