% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getSpecies,dgeComparison-method}
\alias{getSpecies,dgeComparison-method}
\title{Get the species.}
\usage{
\S4method{getSpecies}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with the species name
}
\description{
One of "mouse", "rat", or "human". The last is not yet implemented.
}
\examples{
getSpecies(dge)
}
