% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getSpecies,dgeExperiment-method}
\alias{getSpecies,dgeExperiment-method}
\title{The species.}
\usage{
\S4method{getSpecies}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
The species is returned
}
\description{
The species.
}
