% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{getTimepoint,dgeComparison-method}
\alias{getTimepoint,dgeComparison-method}
\title{Get the timepoint for this comparison. If no timepoint, "None" is returned.}
\usage{
\S4method{getTimepoint}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
Character with the timepoint name
}
\description{
Get the timepoint for this comparison. If no timepoint, "None" is returned.
}
\examples{
getTimepoint(dge)
}
