% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getTimepoint,dgeExperiment-method}
\alias{getTimepoint,dgeExperiment-method}
\title{Get the timepoints in the experiment}
\usage{
\S4method{getTimepoint}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
A character vector of the timepoints. "None" is returned if there are none.
}
\description{
Get the timepoints in the experiment
}
\examples{
getTreatments(dgeExperiment)
}
