% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{getTotalN,dgeExperiment-method}
\alias{getTotalN,dgeExperiment-method}
\title{Get total N}
\usage{
\S4method{getTotalN}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment}}
}
\value{
The total number of animals in the experiment.
}
\description{
Get total N
}
\examples{
getTotalN(dgeExperiment)
}
