% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{heatmap.3}
\alias{heatmap.3}
\title{Plots an interactive heatmap suitable for webpages}
\usage{
heatmap.3(hData, estimate = "hedgesG", ids = NULL, Rowv = TRUE, Colv = TRUE)
}
\arguments{
\item{hData}{The \code{data.table} from which to plot the heatmap.
All the data points in this object will be plotted, so any subsetting must happen before.}

\item{estimate}{The estimate which should be plotted. Must be one of \code{hedgesG} (default),
\code{cohensD} or \code{logFC}.}

\item{ids}{The ids which will be printed. Can be a character vector or data.table
with the ids as well as the compound symbol}

\item{Rowv}{Should the rows be reordered? Defualt is \code{TRUE}.}

\item{Colv}{Should the columns be reordered? Defualt is \code{TRUE}.}
}
\value{
An object of class heatmapr
}
\description{
Plots an interactive heatmap suitable for webpages
}
