% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeComparison.R
\name{printComparison,dgeComparison-method}
\alias{printComparison,dgeComparison-method}
\title{Print the comparison, namely Group1 - Group2}
\usage{
\S4method{printComparison}{dgeComparison}(o)
}
\arguments{
\item{o}{An object of class \code{dgeComparison}}
}
\value{
String with the specified comparison
}
\description{
Print the comparison, namely Group1 - Group2
}
\examples{
printComparison(dge)
}
