% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeData.R
\name{printComparison,dgeData-method}
\alias{printComparison,dgeData-method}
\title{Print the comparisons in this data object.}
\usage{
\S4method{printComparison}{dgeData}(o)
}
\arguments{
\item{o}{An object of class \code{dgeData} for which to print the comparisons}
}
\value{
A vector of strings with the comparisons. The vector is named by the ids of the comparisons
}
\description{
Print the comparisons in this data object.
}
