% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExperiment.R
\name{printComparison,dgeExperiment-method}
\alias{printComparison,dgeExperiment-method}
\title{Print the comparisons in this experiment object.}
\usage{
\S4method{printComparison}{dgeExperiment}(o)
}
\arguments{
\item{o}{An object of class \code{dgeExperiment} for which to print the comparisons}
}
\value{
A vector of strings with the comparisons
}
\description{
Print the comparisons in this experiment object.
}
