% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{smoothForest}
\alias{smoothForest}
\title{Title}
\usage{
smoothForest(
  hData,
  includeModel = FALSE,
  includeName = FALSE,
  includeRegion = FALSE,
  includeEstimate = FALSE,
  includeTreatment = FALSE,
  includeTimepoint = FALSE,
  includeSex = FALSE,
  includePval = FALSE,
  comparisonNames = NULL,
  orderBy = "estimate",
  estimate = "hedgesG",
  splitBy = "gene",
  fontSize = 3,
  ids
)
}
\arguments{
\item{hData}{Needs to be a \code{data.table} object.
The data can contain information for a single gene or multiple genes.
However many genes there are, those many plots will be returned.
The function expects the columns to be the same as those returned by the
\code{flattenDge} method.}

\item{includeModel}{logical. Includes the model if \code{TRUE}.}

\item{includeName}{logical. Includes the experiment name if \code{TRUE}.}

\item{includeRegion}{logical. Includes the brain region if \code{TRUE}.}

\item{includeEstimate}{logical. Includes the estimate if \code{TRUE}.}

\item{includeTreatment}{logical. Includes the treatment if \code{TRUE}.}

\item{includeTimepoint}{logical. Includes the ttimepoint if \code{TRUE}.}

\item{includeSex}{logical. Includes the sex if \code{TRUE}.}

\item{includePval}{logical. Includes the pvalue if \code{TRUE}.}

\item{comparisonNames}{A character vector. If specified, the graph will show these as the comparison names.}

\item{orderBy}{One of \code{asis}, \code{estimate}, \code{experiment}, \code{pval}.}

\item{estimate}{Which estimate should be plotted. Options are \code{hedgesG} (the default), \code{cohensD}, \code{logFC}.}

\item{splitBy}{Generate plots either by "gene" or by "comparison".}

\item{fontSize}{The font size when printing. Defaults to 3.}

\item{ids}{The data.table of mappings between ids and symbols for genes}
}
\value{
A list of ggplot2 objects which can be plotted.
}
\description{
Title
}
